requirements.txt
Flask~=3.1.0
Flask-SQLAlchemy==3.1.1
Werkzeug~=3.1.3
flask-swagger-ui==4.11.1
selenium~=4.29.0
allure-python-commons~=2.13.5
pytest~=7.3.1
requests~=2.32.3
jsonschema~=4.23.0
psycopg2-binary~=2.9.10
Faker~=37.0.0
webdriver-manager~=4.0.2